Clean refactor of code :+1: builds and publishes data to topic
